py multibody: Move non-deprecated code out of `multibody_tree`
