+++ bindings/pydrake/multibody/BUILD.bazel
@@ -47,6 +47,54 @@ drake_pybind_library(
 )
 
 drake_pybind_library(
+    name = "math_py",
+    cc_deps = [
+        "//bindings/pydrake:documentation_pybind",
+    ],
+    cc_srcs = ["math_py.cc"],
+    package_info = PACKAGE_INFO,
+    py_deps = [":module_py"],
+)
+
+drake_pybind_library(
+    name = "tree_py",
+    cc_deps = [
+        "//bindings/pydrake:documentation_pybind",
+        "//bindings/pydrake/common:deprecation_pybind",
+        "//bindings/pydrake/common:eigen_geometry_pybind",
+        "//bindings/pydrake/common:type_safe_index_pybind",
+    ],
+    cc_srcs = ["tree_py.cc"],
+    package_info = PACKAGE_INFO,
+    py_deps = [
+        ":math_py",
+        ":module_py",
+        "//bindings/pydrake/common:eigen_geometry_py",
+        "//bindings/pydrake/systems:framework_py",
+    ],
+)
+
+drake_pybind_library(
+    name = "plant_py",
+    cc_deps = [
+        "//bindings/pydrake:documentation_pybind",
+        "//bindings/pydrake/common:deprecation_pybind",
+        "//bindings/pydrake/common:drake_optional_pybind",
+        "//bindings/pydrake/common:eigen_geometry_pybind",
+        "//bindings/pydrake/common:type_safe_index_pybind",
+        "//bindings/pydrake/systems:systems_pybind",
+    ],
+    cc_srcs = ["plant_py.cc"],
+    package_info = PACKAGE_INFO,
+    py_deps = [
+        ":math_py",
+        ":module_py",
+        ":tree_py",
+        "//bindings/pydrake:geometry_py",
+    ],
+)
+
+drake_pybind_library(
     name = "multibody_tree_py",
     cc_deps = [
         "//bindings/pydrake:documentation_pybind",
@@ -59,16 +107,11 @@ drake_pybind_library(
     cc_srcs = ["multibody_tree_py.cc"],
     package_info = PACKAGE_INFO,
     py_deps = [
+        ":math_py",
         ":module_py",
         ":parsing_py",
-        "//bindings/pydrake:geometry_py",
-        "//bindings/pydrake/systems:framework_py",
-        "//bindings/pydrake/common:eigen_geometry_py",
-    ],
-    py_srcs = [
-        "math.py",
-        "plant.py",
-        "tree.py",
+        ":plant_py",
+        ":tree_py",
     ],
 )
 
@@ -115,8 +158,11 @@ drake_py_library_aliases(
 PY_LIBRARIES_WITH_INSTALL = [
     ":benchmarks_py",
     ":inverse_kinematics_py",
+    ":math_py",
     ":multibody_tree_py",
     ":parsing_py",
+    ":plant_py",
+    ":tree_py",
 ]
 
 PY_LIBRARIES = PY_LIBRARIES_ATTIC_ALIASES + [
