--- bindings/pydrake/multibody/math.py
@@ -1,9 +1,0 @@
-"""
-Bindings for multibody math.
-
-Note:
-    The documentation is not yet populated with symbols, but will be soon.
-"""
-# TODO(eric.cousineau): Replace with actual bindings.
-
-from pydrake.multibody.multibody_tree.math import *
