--- bindings/pydrake/multibody/plant.py
@@ -1,9 +1,0 @@
-"""
-Bindings for MultibodyPlant and related classes.
-
-Note:
-    The documentation is not yet populated with symbols, but will be soon.
-"""
-# TODO(eric.cousineau): Replace with actual bindings.
-
-from pydrake.multibody.multibody_tree.multibody_plant import *
