--- bindings/pydrake/multibody/tree.py
@@ -1,14 +1,0 @@
-"""
-Bindings for multibody tree components.
-
-Note:
-    The documentation is not yet populated with symbols, but will be soon.
-"""
-# TODO(eric.cousineau): Replace with actual bindings.
-
-from pydrake.multibody.multibody_tree import *
-
-# Do not import soon-to-be deprecated symbols (#9366).
-del BodyNodeIndex
-del MobilizerIndex
-del MultibodyTree
