+++ bindings/pydrake/test/all_test.py
@@ -116,13 +116,17 @@ class TestAll(unittest.TestCase):
             "MakeAcrobotPlant",
             # - inverse_kinematics
             "InverseKinematics",
-            # - multibody_tree
-            "MultibodyPlant",
+            # - math
             "SpatialVelocity",
+            # - multibody_tree
+            "BodyNodeIndex",
+            "MultibodyTree",
             # - parsing
             "Parser",
             # - parsers
             "PackageMap",
+            # - plant
+            "MultibodyPlant",
             # - rigid_body_plant
             "RigidBodyPlant",
             # - rigid_body_tree
@@ -131,6 +135,8 @@ class TestAll(unittest.TestCase):
             # TODO(eric.cousineau): Avoid collision with `collision.Element`.
             # Import modules, since these names are generic.
             "Element",
+            # - tree
+            "MultibodyForces",
             # perception
             "PointCloud",
             # solvers
